+++ src/main/java/umc/TripPiece/repository/CityRepository.java
@@ -6,7 +6,7 @@ import umc.TripPiece.domain.City;
 import java.math.BigInteger;
 import java.util.List;
 
-public interface CityRepository extends JpaRepository<City, BigInteger> {
+public interface CityRepository extends JpaRepository<City, Long> {
     List<City> findByNameContainingIgnoreCase(String name);
-    List<City> findByCountryId(BigInteger countryId);
+    List<City> findByCountryId(Long countryId);
 }
